add fake auth
