+++ src/Components/App.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
+import PrivateRoute from '/Users/Cierra/Documents/TP/movie-tracker/movie-tracker/src/Components/PrivateRoute.js';
 import { latestMovies } from '../Actions/movieActions';
 import { isAuthenticated } from '/Users/Cierra/Documents/TP/movie-tracker/movie-tracker/src/Actions/TheActionMan.js';
 import { BrowserRouter, Route, withRouter, Switch, Redirect } from 'react-router-dom';
@@ -8,26 +9,6 @@ import Nav from './Nav/Nav';
 import Login from './Login/Login'
 import * as api from '../Helpers/apiCaller'
 import '../styles/App.css';
-
-
-const fakeAuth = {
-  isAuthenticated: false,
-  authenticate(cb){
-    this.isAuthenticated = true
-    setTimeout(cb, 100)
-  },
-  signout(cb){
-    this.isAuthenticated = false;
-    setTimeout(cb, 100)
-  }
-}
-
-const PrivateRoute = ({component: Main, ...rest})=>(
-  <Route {...rest} render={(props)=>(
-    fakeAuth.isAuthenticated == true ? <Main {...props}/> : <Redirect to='/login'/>
-  )} />
-)
-
 
 class App extends Component {
   constructor(props) {
